server/LLMPromptService/requirements.txt
fastapi>=0.68.0
mangum>=0.10.0
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-openai>=0.0.1
langchain-google-genai>=0.0.1
langchain-aws>=0.0.1
anthropic>=0.3.0
boto3>=1.28.0
uvicorn>=0.15.0
python-dotenv>=0.19.0